[proliferate] split automations domain helpers by subdomain
