+++ desktop/src/hooks/automations/derived/use-automation-mode-selection.ts
@@ -4,7 +4,7 @@ import { useShallow } from "zustand/react/shallow";
 import {
   resolveAutomationModeSelection,
   type AutomationModeOverride,
-} from "@/lib/domain/automations/mode-selection";
+} from "@/lib/domain/automations/mode/selection";
 import { launchControlToConfiguredSessionControlValues } from "@/lib/domain/chat/session-controls/session-mode-control";
 import { useUserPreferencesStore } from "@/stores/preferences/user-preferences-store";
