+++ desktop/src/lib/domain/automations/mode/selection.ts
