+++ desktop/src/lib/domain/automations/schedule/presentation.ts
@@ -1,38 +1,7 @@
 import {
-  AUTOMATION_EXECUTION_TARGET_VALUES,
-  AUTOMATION_REASONING_EFFORT_VALUES,
   AUTOMATION_SCHEDULE_PRESETS,
-  type AutomationSupportedAgentKind,
 } from "@/config/automations";
-import { timeForRrule, type AutomationSchedulePresetOrCustom } from "@/lib/domain/automations/schedule";
-
-export const AUTOMATION_EXECUTION_TARGET_OPTIONS = AUTOMATION_EXECUTION_TARGET_VALUES.map((value) => ({
-  value,
-  label: value === "cloud" ? "Cloud" : "Local",
-}));
-
-export const AUTOMATION_AGENT_KIND_LABELS: Record<AutomationSupportedAgentKind, string> = {
-  claude: "Claude",
-  codex: "Codex",
-  gemini: "Gemini",
-};
-
-const AUTOMATION_REASONING_EFFORT_LABELS: Record<
-  typeof AUTOMATION_REASONING_EFFORT_VALUES[number],
-  string
-> = {
-  none: "None",
-  minimal: "Minimal",
-  low: "Low",
-  medium: "Medium",
-  high: "High",
-  xhigh: "Extra high",
-};
-
-export const AUTOMATION_REASONING_EFFORT_OPTIONS = AUTOMATION_REASONING_EFFORT_VALUES.map((value) => ({
-  value,
-  label: AUTOMATION_REASONING_EFFORT_LABELS[value],
-}));
+import { timeForRrule, type AutomationSchedulePresetOrCustom } from "@/lib/domain/automations/schedule/schedule";
 
 const AUTOMATION_SCHEDULE_PRESET_LABELS: Record<
   typeof AUTOMATION_SCHEDULE_PRESETS[number]["value"],
