+++ desktop/src/lib/domain/automations/schedule/schedule.test.ts
